fix: Code duplication for setup function
